refactor: switch to index based
